Added the AIPath and made Enemies
Imported the AIPath plugin and used it to create a flying enemy that follows the player. Also completed the static enemy 1, which shoots a bullet at random intervals and cannot move. Next up, the last enemy, enemy number 2, who's only special thing will be the fact that it throws diagonal bullets